[fidlc] Add dependency edges for const/enum/bits types
Currently, Library::SortDeclarations does not insert dependency edges
for a constant's type or for enum/bits ordinal types. This is not a
problem from a codegen perspective since we do not emit type aliases in
bindings. However, it causes a bug where sometimes size bound constants
from type aliases are used before being resolved.

This CL adds those dependency edges, so that these constants get
resolved before being used in the Library::Compile() declaration_order_
loop. This was chosen instead of simply checking and resolving the
constant in the problematic spot for the following reasons:

1. In the future FIDL might emit the type aliases in bindings. This
   change would become necessary at that point.
2. There is no harm in adding these edges to the graph; something is
   definitely wrong if they create a cycle.

In addition, this CL:

- Updates fidl-dangerous-identifiers/generate.py to use the new gn
  binary location, and makes its C++ output match the formatter.
- Disables a test in fidl.test.dangerous.constants.lower.test.fidl with
  reference to fxb/38124.
- Adds instructions for debugging in CONTRIBUTING.md.

Test: $FUCHSIA_DIR/out/default.zircon/host-x64-linux-clang/obj/system/utest/fidl-compiler/fidl-compiler-test.debug --case declaration_order_test

Bug: 37936
Change-Id: I4f52d1606e33395d8dbbce3675400712cf93502e