[a11y] Restrict Semantic Tree Commit.
This cl does the following:
1. Removes code where during commit semantics manager was deleting
dangling subtrees since this should be handled by semantic provider.
2. Adds logic to check if there are any subtrees after updates are
applied during commit. If there are any nodes/trees found which are not
reachable from Root node, then fail commit, which results in closing the
channel and deleting the semantic tree.
3. Checks if there are any child nodes in the semantic node, which are
not present in the tree. If found fail commit.
4. Adds more unit tests.

Testing:
  * Build: Passed
  * Unit Test:
    * fx run-test a11y_lib_tests (Passed)
      fx run-test a11y_tests (Passed)

BUG=36806

Change-Id: Ib506313cd2eca1a0363ba6a41ac2ae77b0ba8b55