[dhcp] Implement resend timer
This change introduces a resend timer to the DHCP client. Prior to this
change, a client would send a single DISCOVER/REQUEST and then wait for an OFFER/ACK
until it timed out. In a congested environment, the client needs to be
able to resend its last message without timing out the entire
transaction if after a reasonable delay it hasn't received a response.
The resend timer introduced by this change implements this desired
behavior.

Change-Id: Ib75b5eb4263d728737a44c5a3892f66f2ac9c378