[brcmfmac] Process CLM.blob and set country code
If the CLM blob file is found, read and load it into the firmware.
Added the option to add a country code table that is platform specific.
There is also an option to add a list of iovars that is platform
specific.

The country code table is available as part of wifi metadata that can be
retrieved by the driver. The table is searched for the specified country
code and the appropriate iovar is issued to firmware.

If the iovar table is not empty, the specified iovars are issued to
firmware during init.

Test: Checked that loading CLM.blob and setting country code works as
expected in the real hardware. Also verified by running the existing
test - lifecycle_test, which exercises the changed code path passes.

Change-Id: I7109165fa6d6576bfc904abc0159ae4c98b07cd8