[componentmgr] Move deleting children to a special 'deleting' directory in hub
When a child (static or dynamic) is being deleted, it is moved from the
'children' directory to the 'deleting' directory in the hub of its parent.

Corrected a bug in AddableDirectoryWithResult where the DirectoryEntry was
not being returned when remove_node was invoked.

Removed BreakpointEvent in favor of EventType.

Added a new Event 'PreDestroyInstance' that is triggered when the destruction
process begins for any child.

Test: Added to dynamic_child_create_bind_stop_delete_test in hub_integration_test.rs
      Reran all tests in hub_integration_test.rs

Change-Id: Ic83a7087c0544501691c704b34c2051ef8429876