[amlogic][sdmmc] Ignore spurious interrupts
The sdio interrupt thread used to break out of its interrupt loop when
an unexpected interrupt happened. When working on DMA for the brcmfmac
wlan driver it seems that some spurious interrupts occur because of CRC
errors and the sytem cannot recover from this because the interrupt
thread dies.

This change simply changes a break to a continue to keep going. The
brcmfmac driver recovers from this without issue, it will drop a frame
but recover and the system can keep going.

Test: wlan-smoke-test, wlan-stress-test, manual testing
Change-Id: I4d0a77f2199fc65ea13ae293353cc198ce982c6d