[fdio][vx] Remove temporary fdio _mmap_file hack
This change is part of implementing the Verified Execution: Controls on
Executable Memory design. See go/fvx-executable-memory for more details.

This removes a temporary hack in fdio's _mmap_file to support usage of
mmap with PROT_EXEC in Flutter. This is no longer needed thanks to
https://github.com/flutter/engine/pull/12703 and was the last remaining
zx_vmo_replace_as_executable call in fdio.

Bug: 35899 #fixed
Change-Id: Ib3cdda47c782b3ab5015278a87de4ba17a790ac9