[acpi-battery] remove polling thread
With the addition of object signal events for significant changes
(i.e. +/- 1% change in battery level) and given the presence of
the ACPI notify handler, the polling thread is now superfluous.

Note: this change has been verified on chromebooks, but should also
be verified on other hardware where applicable so as to avoid any
regression on a device that may actually depend on polling.

For EC implementations that notify immediately as a result of BST
object evaluation, a rate limit on the notification is set (10ms)

Test: regression check, give that this is just removal
      run unit tests (-- with-base //garnet/tests/zircon:zircon_tests):

      fx shell runtests -t x86-battery-test

Change-Id: I36c40088b5fa461aea79da5044e7d654b1d3d3fb