[debugger] Correctly inject an Arch Provider.
The previous approach had a global ArchProvider, which made testing
awkward. This patch makes it follow an injected dependency approach,
similar to what the ObjectProvider does.

This change will be followed by patches that further inject this
dependency up the "tree" (process and debug agent), as well as clean up
some constructors that are getting unwieldly (mostly DebuggedThread).

TEST=Updated unit tests = fx run-test debug_agent_tests
     Tested sw breakpoints manually.

Change-Id: Ie76c4a7432ba51204e143f999e4548f3d826d4f3