[appmgr] Log warning for uses of deprecated-shell
Enforce an allowlist for "deprecated-shell".
Set as penalty a low severity log message.
Initialize allowlist with enough elements to not cause too much logspam
on day one.

In future changes we'll further populate the allowlist and later
escalate the penalty (log warning, log warning and deny launching).

This change as it is can be used to inform people not to use
deprecated-shell (if they read the logs, which they probably don't),
and to harvest CI/CQ for missing allowlist entries.

Bug: 36501
Bug: 35365
Bug: 36227
Change-Id: Iebaad6267720b40ad215824510ef66e3680ab34e