[nmh][reachability] Adding basic functionality and tests
Adds ability to determine reachability state based on knowledge of existing
network configuration.
The states are not fully correct, as there is no access to neighbor
tables and there is no active probing being done yet.
The aim is just to set up a framework to add functionality and
eventually be able to infer the correct state.

Change-Id: Ide809868ee507d6664ba2f67eda9a3a3a7ce54e2