[zircon] Re-enable -Waddress-of-packed-member
When we take a pointer to a member of a packed struct, dereferencing the
pointer may be dangerous if the pointer is misaligned. This is what the warning
checks for, although it instead seems to just warn on taking the address of a
struct member when it's converted to a regular pointer. Before it was
disabled, but we would like to re-enable it now as a part of the GN/ZN
unification.

Enabling this warning raises a couple of false-positives, although most
of them refer to structs that don't need to be packed. We can remove
instances of this warning appearing by simply removing `__PACKED` from
those struct definitions and static_asserting that they remaing the same
size after the removal.

For other instances (or rather just tcs3400) where we cannot remove
`__PACKED`, we just silence the warning at each instance by casting to a
void* and back. This is a temporary solution until we can improve where
the warning is raised. Near the source location of the warning though
was a potential access to a misaligned pointer that this CL fixes.

Bug: 3203, 37765
Change-Id: Id43e47d544d4debaba77e5ec9504c2caac40ddab