[pm] Make sure to copy a versioned-prefixed root metadata file to amber-files
TUF-1.0 requires that the root metadata be versioned, but pm.gni was
only copying the root-manifest.json to
`$OUT_DIR/amber-files/repository/root.json`. This updates the file to also
copy the current metadata to `6.root.json` to be compliant, with a TODO to
come up with a longer term solution that automatically copies the files
when we rotate the root metadata.

Bug: 38262
Test: Verified 6.root.json ended up in amber-files/
Change-Id: Iaa5f3b635dc9e89f02f8ab307ceb1c15e97df9ef