[componentmgr] Simplify Events lifetime
This CL makes a number of simplifications to the Event object's
lifetime.

The Event object no longer holds references and so it no longer needs a
lifetime parameter. Furthermore, it no longer holds a RealmState and so
we don't need to hold a RealmState lock while calling out to hooks.

Event is also now cloneable because it no longer holds references and so
it can be transferred over a mpsc::channel. This enables a bit of
simplification in HubTestHook and the hub_integration_test. It makes
breakpoints even more extensible because forwarding of new events to
hub_integration_test comes for free, effectively.

Change-Id: I60d939778d73524caaa4b0a2404dab0f0b5fa06a