[debugger] Plumb ArchProvider all the way up.
This permits a way to inject an ArchProvider to the debug agent that
will "trickle" down the hierarchy, permitting tests to inject a test
view of both the architecture and the zircon object hierarhcy.

Re-wrote the DebuggedThread and DebuggedProcess constructors to receive
configuration structs, as if was getting unwieldly and confusing with so
many fields.

Also cleaned up the MockProcess and MockThread constructors.

Removed the ObjectProvider::Get static call, that looked like a
singleton factory. Callers can create their own instances.

TEST=Updated all unit and integration tests.

Change-Id: I73ae4c9c684bbe1983baa6c7b988ada56e12a2b2