[pm] Fix race in `pm serve` between starting sse server and broadcasting events
The `serves /auto and events on timestamp updates` test was rarely flaking
due to a race between starting the sse server for some client, and
broadcasting events to the client:

* The test client creates a request to the `/auto` endpoint. Then it then
  waits for the response.
* Context switch to the server.
* The `/auto` handler receives the request, and it starts an sse server.
* The sse server validates the request, then responds with a 200 and
  flushes the response to the client. It then context switches to another
  goroutine before completing the function.
* The test client sees the 200 and touches `timestamp.json`.
* The server fs watcher sees the file changed, and it tells the `/auto`
  handler to broadcast the file changed. The `/auto` handler currently
  has no waiting clients, so the event is ignored.
* Context switch back to the `/auto` handler. With the sse server started,
  we add the client response writer to the map of clients wanting to
  observe events.
* Context switch back to the client, and it blocks forever waiting for
  the sse server to send it an event.

This fixes the race by aquiring the AutoServer lock before we start the
sse server, which makes sure we block broadcasting events until clients
have been fully registered.

Test: There isn't a particularly good way to build a test case around this
bug to verify this bug has been fixed, because it's relying on the
scheduler to context switch at an inopportune time. This patch was validated
to fix the bug by continously running the test while the machine was under
extreme load in order to increase the odds of reproducing the bad context
switching order. With this fix applied, the bug didn't show up after more
continuous testing.

Bug: 37107
Change-Id: I82969e09790bc58033b4a2f6eb3a97295101addf