[kernel][vm] Provide option to limit ASLR entropy
Instead of picking an allocation uniformly from all possible locations
this change puts a cap on the number of locations that shall be
considered.

The default entropy for now is 36, which results in emulating the
previous uniform allocation strategy, but this is expected to be
changed in the near future.

Bug: 34546

Change-Id: Ia7470c402f1354a9ed9b9aa0da2b5bf8bdfa4197