[hid-parser] Fix lifetime issues with FidlReport
Store the report_data in the FidlReport so
the FidlReport no longer needs to have the
same lifetime. If these two structs were
required to have the same lifetime, it's
likely this could introduce bugs in the
future. We can avoid this by making a copy
of the data and storing it in the FidlReport.

Also update SetMouseReport and SetMouseDescriptor
to be generic so they can be used for more devices
in the future.

TEST: No functional changes, hid-input-report-test
      covers this API.

Change-Id: Ib6b44dd5da538d54e1f0ffa3c828b39634304066