[fuzzing][build] Don't instrument libc.
This CL adds a new config, similar to `no_sanitizer`, called
`no_fuzzer`, which disables the fuzzer instrumentation for a given
target, by setting the `-fno-sanitize-fuzzer` compiler flag.

The CL also sets the config for the following targets: musl, scudo,
zx-panic-libc, ldmsg. So that libc.so doesn't contain any fuzzer
instrumentation.

In particular, we are interested in removing the code coverage
instrumentation added by `-fsanitize=fuzzer`. While running a fuzzer,
libfuzzer sometimes makes call to the libc, adding noise to the coverage
information, and mistakenly marking corpus as interesting.

This change attempts to solve this issue, together with CL 327491,
disable the instrumentation of the two libraries that are used from
libfuzzer: libfdio.so and libc.so. Sadly, the approach taken for
disabling instrumentation in libfdio cannot be done for libc without
a lot of additional work.

This change is brittle and will break in the future whenever new
dependencies are added to the libc or one of its dependencies. But, it
is relatively easy to find and fix the offending code.

BUG=25073
TEST= Check that the coverage of noop fuzzer is always 1 and that
there is only one module with 1 PC.
```
$ fx set core.x64 \
    --with-base //garnet/tests/zircon:zircon_fuzzers \
    --with-base //bundles:tools \
    --with-base //bundles:tests \
    --with-base //examples/fuzzer \
    --fuzz-with asan
$ fx run -k -N
$ fx fuzz start -f zircon_fuzzers/noop-fuzzer.asan
```

Change-Id: I614bb1867e146fdf3a311a77b09bb4e08576a2a8