[media][tests] set VAD for data passthru
This CL relates to audio test suites' usage of Virtual Audio
Devices (VAD).

Test binaries audio_admin_tests and audio_loopback_tests use
audio_core's loopback capture functionality, which is only
enabled when an output device is present. These tests create
a virtual audio output device, to guarantee that a loopback
capture stream can be created.

When looping-back audio in this manner, one can set the
formats for Renderer and Capturer to identical values, but
this isn't sufficient to avoid potential frame-rate and/or
rechannelization conversion. Additionally, the output device
must be using that same format.

This test-only CL configures the virtual audio output device
at the same format as the renderer and capturer, so that
any possible channel- or rate-conversion is avoided. This
will eliminate possible failure modes, as new resamplers are
added.

Test: validated the change by executing the test binaries --
 fx run-test audio_admin_tests
 fx run-test audio_loopback_tests

Change-Id: Id01b38b8a0c35c6ff7341a84daded4a13f7a18db