[componentmgr] Use fuchsia.sys.Loader when v1 component
When component_manager is run as a v1 component, the fuchsia-pkg
resolver now delegates to `fuchsia.sys.Loader` instead of
`fuchsia.pkg.PackageResolver` to resolve packages.

This change allows component resolution from the base package set to
succeed even if the TUF server is unavailable or the network is not up
yet. Previously, this required the TUF server to be running and
accessible or for the package resolver to be mocked out. For example,
this allows a component_manager registered as an app in a sysmgr config
file to resolve components from the base package set regardless of
component startup order.

This change works because `fuchsia.sys.Loader` implements fallback
behavior to load from pkgfs when pkg_resolver returns NOT_FOUND.

This change also removes mock_pkg_resolver because it's not longer
necessary (hermetic test environments still receive fuchsia.sys.Loader).

Bug: 34844

Change-Id: I1f33d1e4601e92463d19851a82b09d6aa7f987b6