[debugger] Connect to agent pre-connected processes.
If the client connected to an already running agent, there was no way to
attach to those processes, as the agent would return
ZX_ERR_ALREADY_BOUND at an attach request to an already attached process
(it still does).

Now the client detects that situation and asks for the process status,
which makes the agent re-send the notifications. This makes the client
attach correctly.

TEST=Added unit test.
     Tested manually:
     1. fx shell -- run debug_agent --port=5000
     2. fx debug --port=5000
     3. [zxdb] attach <some process>
     4. [zxdb] set quit-agent-on-exit false
     5  <quit zxdb>
     6. fx debug --port=5000  // This is the re-connection.
     7. attach <pid>
     8. ???
     9. Profit!

Change-Id: Id443400ea1f27740002e5d0102a9f57819eac53f