[build] Allow zircon host tests to specify runtime dependencies
This change will allow zircon tests to specify files that it needs at
test-time. These need to be captured in some machine-readable way, as
the infrastructure does not run host tests out of the build directory,
but rather on a bot divorced from the checkout. This mirrors what was
done for the fuchsia build.

In order to use this feature for a given host test(), the following
needs to be the case:
(1) runtime dependencies need to be reachable in the metadata graph of
the test target and use the data key of "test_runtime_deps".
(2) all such dependencies must be found in the build directory
(3) a test must refer to its dependencies via a relative path from the
executable.

Note, if one wishes to use a dependency from the checkout not generated
from the build, one can merely copy() it to the build directory.

Example:

```
group("my-host-test") {
  deps = [":my-test($host_toolchain)"]
}

test("my-test") {
  # the usual...
  deps = [":my-test-runtime-deps"]
}

group("my-test-runtime-deps") {
  deps = [":source-deps"]

  metadata = {
    test_runtime_deps = [
        "$root_out_dir/foo.txt",
        "$root_build_dir/bar.json",
    ]
  }
}

copy("source-deps") {
  sources = ["//src/baz.txt"]
  outputs = ["$root_out_dir/baz.txt"]
  metadata = {
    test_runtime_deps = outputs
  }
}
```

Bug: 37774
Change-Id: I1ca8bf837a8d1e8e834c21d99f5eb0156260e7ac