[Setting] Send ACK to VolumeControl
Volume Control needs to receive an ACK that volume
changes are being handled by the client. If it
doesn't receive the ACK for a max number of events,
then it'll disconnect the client.

This caused an issue where if the audio settings
cannot change the volume past the non-ACK threshold.
The CL fixes the issue by listening for volume
changes and then send an ACK to Volume Control.

Bug: fxb/37777
Test:
The tests in audio_tests are modified so that they
would also check for acks.
You can run the tests via:
   fx run-test setui_service_tests -- --test

Change-Id: Id2bc9cbacc59dd76cb3243de61dc89ad2fec2859