[fbl][test] Convert function tests to zxtest
Also, move the compare-function-to-nullptr operator overloads out of the fbl
namespace and into the global namespace.  It is quite difficult to invoke an
overloaded operator for an object when it is not in your namespace, and I'm
pretty sure that the original intention was to make it easy to compare a
function container object against nullptr.

Change-Id: Idea64d7ec020ba7af7e0caecaff3c1911e71e750