[debugger] Don't assume the unit when decoding DIEs.
Previously the DwarfDieDecoder took the compilation unit as a
constructor parameter. This allows certain things to be cached.

But it's possible (though uncommon) for abstract origins to cross
compilation unit boundaries. Caching the unit in this case will be
incorrect. This converts the callers to always take a DWARFDie (which
includes the unit) and avoid any unit-specific structure members.

TEST: All of this code is exercised by the existing tests. The thing
that's not tested is the case where this new code works and the old code
doesn't. I've only ever seen that once in a 167MB binary and I'm working
on some infrastructure to be able to put this somewhere for tests
(issue 38236).

Change-Id: I39c3e5ccc3fcdc961639531dfe1de7cee89086af