[oom] Remove GN arg used to configure OOM behavior
There was previously no way to pass kernel command line args depending
on product config, so a different kernel was built to configure the OOM
behavior. This is now plumbed through, and it is preferable to have the
same kernel build configured by command line, so instead add
'kernel.oom.behavior' to control whether the system should attempt the
job killing, or (default) just reboot.

No intended change in behavior.

Bug: 33429
Bug: 32577
Bug: 33428
Bug: 28456
Change-Id: Ib058fa2e977187eab32ea1acef7dc51b6ba4ebbd