[debugger] Improve breakpoint messaging.
Add a message when a breakpoint is resolved. This helps pending
breakpoint seem less mysterious. A breakpoint observer function is added
for this case. There was an existing observer function that was never
called which was removed.

Update the breakpoint observers to use the new centrally-managed style
on the Session. This saves the ConsoleContext from having to register
for every single object.

Add the number of matched locations to breakoint descriptions.

Add a test for descriptions and some mocking to make this practical.

Add an error message for the "not supported" return value when setting a
breakpoint to indicate the new kernel setting.

Change-Id: I1c2697d5e5e3d9e7e9a0cc66c13d3954e41f68b8