[netemul] Query daemon state before tests.
The functionality of the serial port has been used to
signal that tests are ready to run.  This change adds an
explicit check of systemctl to ensure that the
guest_interaction_daemon is running before starting tests.
This ensures that there is something listening on the vsock
port for the GuestInteractionService to connect to.

Bug: 38161
Change-Id: I94e83a833ea56883cf693a8e1de78eb44dec5a0e