[debugger] Use permissive input locations
Hooks up the remaining input location parsing to use the new
"permissive" input locations. These will match items in any namespace.
This is mostly a cleanup from the previuos work.

Removes an unneeded Resolve* variant.

Updates the tests since the new infrastructure require the symbol be
properly in the index.

Change-Id: Id8c2ac4fbcdc18a096b6ebe458572cf4df8070f7