[fidlgen][rust] Add support for strict xunions
Currently, the Rust backend treats `strict xunion` the same as `xunion`.
This CL teaches it to use strict semantics. This means (1) the generated
enum has no unknown variant, and (2) decoding fails in the bindings with
Error::UnknownUnionTag if it encounters an unknown ordinal. In addition,
the enums for strict xunions are allowed to derive the Copy trait
(flexible xunions cannot because the unknown variant contains vectors).

This CL splits the fidl_xunion macro into two new macros:
fidl_flexible_xunion (same as the old one) and fidl_strict_xunion
(modified to have strict semantics).

Test: fx run-test rust_fidl_tests # new unit tests in encoding.rs

Bug: 34780
Bug: 38186
Change-Id: Iad96b0386b3095e6a55ecc1abfe0bbd563feebd9