[blackout] add random directory tree generation
Generate random directory trees to be written to disk as part of the
tests, putting more load on the filesystem. This is closer to some kind
of install than normal operation - we just commit that entire tree to
disk, never removing it.

In empirical testing, we haven't run out of space when writing trees to
disk. If we do, the host-side harness will mark it as a non-verification
error and move on.

Bug: 34446
Test: fx run-test blackout-target-test
Change-Id: I28d5ead99b185eda7fdfacd37d26be968451ecbb