[codec] FakeMapRange utility
This will be used in a later CL that'll add zero-copy of decoder input
buffers.

The point of this utility is to avoid mapping secure buffers, but keep
handling of secure and non-secure cases similar.  In particular, we'll
continue passing around buffer base()-relative virtual addresses.  This
is mainly because we have code that _does_ need to touch the buffer data
with the CPU, but that code happens to not be used in secure cases.

With this we'll be able to robustly fault on any attempt to touch secure
buffers with the CPU, without running into the strange behaviors that
can occur if one actually tries to touch a mapped secure buffer with the
CPU.

We'll share one FakeMapRange across all buffers of a secure
BufferCollection.

Change-Id: I1f115706aba271291c7273a9f2af6c1d2e1f137f