[bt][a2dp] take_transport Result, add AsyncWrite
Change StreamEndpoint::take_transport() to return a Result instead of
panic!ing when the transport is already taken.  Update bt-a2dp-sink as
appropriate.

Add tests to check this behavior and possible re-acquisition of the
MediaStream if it is dropped.

Implement futures::io::AsyncWrite for MediaStream in preparation for
A2DP source, and add tests to confirm it works and also behaves
appropriately when the underlying channel disappears unexpectectedly.

Bug: 1221
Test: added test_mediastream and test_asyncwrite
Change-Id: I5d6399b6b4d68653d41603621f8003150469d4d2