[virtcon] Reland "[virtcon] Convert to use sysmem""
Needed bug fix fxr/327498 and new test fxr/327674

Original description:

In the future we want to remove the legacy ComputeLinearImageStride and
AllocateVmo calls on the display controller, since they're redundant
with sysmem's memory allocation system (and not flexible enough to work
with things like vulkan). virtcon is one of the last pieces of code
using those calls, so switch it over to using sysmem.

Test: vc-display-test
Bug: 33157
Change-Id: I01d7986071d4b8824a11e4dd855f123065d8ef04