[rust][third-party] Add extra-traits feature to syn
This is necessary to ensure that other traits such as
Debug are implemented by syntax tree types. This
feature was being enabled by other third-party targets,
but was unavailable when building with fargo or other
tools based on the generated Cargo.toml files.

This fixes an issue that was blocking the Rust docs roller.

Change-Id: I730c101cb514d46307945eafadbae644bcb2b18c