[kernel][x86] fix bug in cache topography reading from cpuid
When reading cpuid 0x4 and 0x8000001d two things were going wrong.
First, it wasn't checking to see if it was a valid cpuid leaf and on
some (all?) intel cpus a read from a bad cpuid returns the last valid
leaf. This was confusing the cache reading code.

Second, there was a logic error that caused it to not actually pick up
the highest level cache leaf as it went through the loop looking for the
highest one set.

This was causing it to not actually pick up the last level cache
structure and pass through to the topology generator.

Also cleaned up a few printfs here or there to print slightly more
information. Most of them are LTRACEFs only so little change

Tested: Booted on Intel Core i7-6100 and AMD Threadripper 1950x and
qemu. Validate that the proper cache hierarchy is picked up. Validate 'k ut
x86_topology' still passes.

Change-Id: I6f1ffb78fd38c1cfadc740698305f2bb5274db93