[netstack3] Disable DAD during tests
Disable duplicate address detection during the netstack3 integration
tests are running. This is needed to disable the delay in registering IP
addresses. This small delay was causing future tests to fail due to
assuming the IP address is set immediately.

This should be replaced when an event is exposed outside the netstack
for when duplicate address detection finishes. With such an event, any
client can confirm whether the IP address was successfully set or not.

Bug: 36238 tracks the above TODO
Bug: 21542
Change-Id: I4d69691cf3b68f833d695bf7777769c1fbe673f0