[bt][a2dp] Add a config which starts a2dp-sink on device startup
Add a bluetooth config_data which starts bt-a2dp-sink on device startup.
This will have the side effect of starting bt-init and the rest of the
core bluetooth stack, which bt-a2dp-sink requests.

In order to use this, required components for both media and bluetooth
will need to be in the base image, as these apps may be run before the
package server is active and able to download ephemeral packages.

e.g. to include via fx set:

  ```
  $ fx set
   --with //src/connectivity/bluetooth/profiles/bt-a2dp-sink:startup_config
   --with-base //src/connectivity/bluetooth
   --with-base //garnet/packages/prod:media_audio
   --with-base //garnet/packages/prod:mediaplayer
  ```

Bug: 36461
Test: 1. Build fuchsia with //src/connectivity/bluetooth/profiles/bt-a2dp-sink:startup_config
      2. Run `ps` and verify that bt-a2dp-sink.cmx and all of its
      dependent components are running (bt-init.cmx, bt-gap.cmx, bt-snoop.cmx, mediaplayer.cmx,
      audio_core.cmx).
      3. Use bt-cli and bt-le-central to verify that the Bluetooth
      system services are functional.
      4. Use bt-cli to make the device discoverable, pair, and verify
      that streaming music works.

Change-Id: Icef07852a7bb7376e7496bdc9575804037541f76