[debugger] Add cross unit index test.
Tests a case of cross-compilation-unit indexing that was fixed by the
previous couple of patches but not fully tested.

Since this requires a large amount of test data, the data is not
downloaded and the test is not compiled by default. The team can enable
this to manually run this new class of "large" tests.

Do some various build improvements. The gtest_prod target is now
available, and the copy_sample_core step should list all the files for
proper dependency tracking.

BUG=38236

Change-Id: I635226d5e5ee44b645f0225b35e1e8134ebb1f4b