[net-cli] Propagate errors and use logger to log
Errors returned from Netstack as a result of FIDL methods will be
propagated up and cause the binary to exit with an error.

Instead of using println!() and eprintln!() directly for output, added a
logger which prints info or below to stdout and warn or above to stderr.
Output indicating an operation completed successfully is now logged at
info, which by default will be shown; while output that should always be
shown (e.g. interface information when net-cli is invoked to list all
interfaces) still use println!().

Test: fx run-test net-cli-tests
Change-Id: I2470db5d3dc107f7f6a39e2e14c720e938479f99