[mediaplayer] accommodate decryptor stream processors
The StreamProcessor node type, which is intended to handle decryptors
in addition to its current job of handling decoders, assumed that the
processor in question would know its input stream type in advance.
This is true of decoders, whose creation takes a stream type as an
argument.

Decryptors will be created by the client based on a parameter block
that does not include the input stream type. This means that
StreamProcessor nodes that are used to wrap decryptors need to
be configured with the input stream type well after they are
created. This CL added a method to stream processors to enable
this.

TEST: added unit test for new SetInputStreamType method

Change-Id: Ia1dcc3f85307eed5c3fd8c611d85da7d4b21e70f