[scenic] Promote ViewHolder from Resource to Node
ViewHolder is already a node from Scenic's perspective. Exposing this aspect
on the SDK lib to be consistent. This API change is backwards compatible since
Node is a subclass of Resource.

Altered a pixel test and the input coordinate test to exercise newly exposed
functionality, and added a detach pixel test for fxb/24510

fxb/24510 #done

Test: fx run-test scenic_tests
Change-Id: Ibd7619d2036205a5cf626d3c836490420b3559d1