[debugger] Moved SW breakpoints specific concepts.
Before this CL the logic of stepping over a SW breakpoint was part of
ProcessBreakpoint. Now that ProcessBreakpoint is an interface (HW
breakpoints implement them too), that logic has to be moved to the
SoftwareBreakpoint implementation.

Defined the virtual interface all breakpoints have to implement.
Hardware breakpoint one is stubbed for now.

TEST=No functional change. All tests passes.
     fx run-test debug_agent_tests

Change-Id: Ia8e58f4a5d4b90d6b70835c31776ee036c82d818