[media][audio_core] Use zx::time for driver timeouts.
Also fixes a bug in the command timeout calculations. Additionally we
make some changes to the FakeAudioDriver and MessageTransceiver that
allow us to control the flow of the fake driver. We also remove one of
the async waits in the MessageTransceiver that's not needed since
zx_channel_write will never return ZX_ERR_SHOULD_WAIT.

MULTIPLY: `[{
    "target": "//src/media/audio/audio_core/test:audio_device_service_tests",
    "total_runs": 100
}]`

Test: fx run-test audio_core_unittests
      Added additional test case for command timeout.
Change-Id: Ida5fb65f14c97bf751a5873475153bd896ee7e58