[coroutine] Don't reuse coroutines in CoroutineManager
Once coroutines have been resumed with ContinuationStatus::INTERRUPTED,
they cannot yield anymore (any attempt to Yield will DCHECK). The
coroutine manager was trying to reuse its coroutines as long as it is
not destructed, but coroutines can also be interrupted if they do a
SyncCall and the callback is dropped.
Instead of reusing coroutines, we rely on the coroutine service reusing
stacks. This leads to (temporarily) using one more stack: when one
pending task completes, we need to launch the next task in another
coroutine before returning from the current one.

Test: CoroutineManagerTest.MultipleCoroutinesInterrupted
Change-Id: I1abdb02de44565fe2570d937df1da6235b7393ad