[bt][lib] Represent assigned numbers as integers
The current string representation maps well to the existing FIDL APIs
that represent UUIDs as strings however don't work well with the
concrete types. Rather than constructing types::Uuid instances from a
string it is more natural to construct them from a 16-bit integer.

The AssignedNumber type and constant list now represents the "number"
value in the 16-bit representation. The match functions now construct a
types::Uuid to convert these into the full 128-bit form.

Updated the bt-le-peripheral tool (which uses the AssignedNumbers)
utility to represent all UUIDs as types::Uuid and convert them to a
string at the point of use. This lends itself well to the new API, in
which the tool will convert these to the FIDL type using the existing
conversion functionality.

The tool now also treats garbage UUID strings as an error rather than
passing them down as-is to the FIDL API.

Bug: 1405
Test: 1. bluetooth-crate-unittests
      2. Run bt-le-peripheral with various "-s" combinations:
         a. "-s GATT" completes to the 128-bit form of 0x1801
	 b. "-s aaaabbbb-cccc-dddd-eeee-ffeeddccbbaa" uses the UUID as
	 is
	 c. "-s poop" prints a message "invalid UUID: poop"
Change-Id: Iec158153c7166a65a8e953c3a3bd08400f86cf1d