[futex][locking] Break up the futex context lock.
Instead of having a single lock which protects all futex operations for an
entire process, break the lock up in to a process-wide "pool", and an
individual kernel mutex for each FutexState.  The pool lock never performs _any_
potentially blocking operations and therefor may be a spin lock instead of
needing to be a heavier kernel mutex.

The individual state locks need to stay as kernel mutexes, because state
validation may result in a page fault which needs to be serviced by a user mode
page fault handler.  Now, however, blocking like this can only block users of
that individual futex, and not all of the futexes in the process.

The price of this lock decomposition, however, is that each thread now needs to
contribute two FutexStates to the process-wide pool instead of just one.  See
the comments in FutexRequeue for details, but the short answer is that it is
Requeue's fault.

Test: Existing unit tests and long term stress testing runs.
Change-Id: I0faa5309c9a35d18595214b4ace94d59a26f9445