[netstack2] Fix endpoint double close bug and handle receiving TCP RST immediately
This change addresses a bug which leads to an endpoint (that can have references
to it by multiple socketImpls) may attempt to cleanup its resources multiple
times, resulting in a panic when we attempt to close a closed go channel.

The bug:
1) Accept is called on a listening socket.
2) Create an endpoint and a socketImpl referencing the endpoint, return the
   socketImpl.
3) Receive a TCP RST which closes the endpoint (it only had a ref count of 1
   which drops to 0 so it gets closed).
4) Clone the socketImpl which results in a new socketImpl but with a reference
   to the original endpoint (the ref count of the endpoint increases to 1).
5) Close the new socketImpl (which closes the endpoint again since the ref count
   drops back to 0).

Step 5 is where the bug occurs (before this change, the cleanup would would be
done twice).

This change also addresses another bug where a TCP connection may be closed
immediately after being accepted (receiving a TCP RST on an endpoint that was
just Accepted but not yet returned to the caller of Accept).

Test: Test that closing an endpoint twice will not result in a panic.

Bug: 37475

Change-Id: Ie19f45cb2d864ee13f64ae1685eaa205a3802a46