[blobfs] Add trim support.
Blobfs now issues TRIM commands to block devices that advertise
support for the feature.

TRIM requests go through the journal queue but they are not journaled.
In other words, a crash after the command is issued will not result in
replay reissuing the request, so it is possibe for the filesystem
to consider the deletion "done" without the block device seeing it.
This is OK as TRIM is just and advisory command.

Bug: 31379

Change-Id: I6aa4785a7fdc7bab16091afa0881cfc53e8b7ec4